Bug fix
When updating a cards and switching the input to file, before exiting. the input will stay a file input. therefore when you try to update again and the program tries to enter the current image into the text input, it stops because it's now a file input.

Problem was fixed by switching the input back to a text type in case it's a file type